lab 1 + inmultire si impartire

+++ isp-lab-1-2023-master/src/test/java/lab/template/ExampleTest.java
@@ -16,11 +16,18 @@ public class ExampleTest {
     }
 
     @Test
-    public void testSubtract() {assertEquals("Should subtract 2 numbers", 5, Example.subtract(15, 10)); }
+    public void testSubtract() {
+        assertEquals("Should subtract 2 numbers", 5, Example.subtract(15, 10));
+    }
 
     @Test
-    public void testMultiply() { assertEquals("Should multiply 2 numbers", 150, Example.multiply(15,10)); }
+    public void testMultiply() {
+        assertEquals("Should multiply 2 numbers", 150, Example.multiply(15, 10));
+    }
+
     @Test
-    public void testDivide() { assertEquals("Should divide 2 numbers", 1.5, Example.divide(15.0F, (float) 10.0F), 0.001); }
+    public void testDivide() {
+        assertEquals("Should divide 2 numbers", 1.5, Example.divide(15.0F, 10.0F), 0.001);
+    }
 }
 //implementati inmultire si impartire
